Extract out all markdown-like commands
