--- lib/streamlit/elements/latex.py
@@ -1,42 +1,0 @@
-from streamlit.proto import Markdown_pb2
-from streamlit import type_util
-
-
-class LatexMixin:
-    def latex(self, element, body):
-        # This docstring needs to be "raw" because of the backslashes in the
-        # example below.
-        r"""Display mathematical expressions formatted as LaTeX.
-
-        Supported LaTeX functions are listed at
-        https://katex.org/docs/supported.html.
-
-        Parameters
-        ----------
-        body : str or SymPy expression
-            The string or SymPy expression to display as LaTeX. If str, it's
-            a good idea to use raw Python strings since LaTeX uses backslashes
-            a lot.
-
-
-        Example
-        -------
-        >>> st.latex(r'''
-        ...     a + ar + a r^2 + a r^3 + \cdots + a r^{n-1} =
-        ...     \sum_{k=0}^{n-1} ar^k =
-        ...     a \left(\frac{1-r^{n}}{1-r}\right)
-        ...     ''')
-
-        .. output::
-           https://share.streamlit.io/0.50.0-td2L/index.html?id=NJFsy6NbGTsH2RF9W6ioQ4
-           height: 75px
-
-        """
-        if type_util.is_sympy_expession(body):
-            import sympy
-
-            body = sympy.latex(body)
-
-        latex_proto = Markdown_pb2()
-        latex_proto.body = "$$\n%s\n$$" % _clean_text(body)
-        dg._enqueue("latex", latex_proto)
